move ErrorCollector to own file
